requirements.txt
numpy>=1.21.0
opencv-python>=4.5.4
pytesseract>=0.3.10
Pillow>=9.0.0
pandas>=1.3.5
scikit-learn>=1.0.2
matplotlib>=3.5.1
seaborn>=0.11.2
python-dotenv>=0.19.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.13.0
openai>=1.3.0
requests>=2.28.0
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.1.0
flake8>=6.0.0
mypy>=1.0.0
